+++ Client/N3Base/N3UITrackBar.h
@@ -24,9 +24,9 @@ protected:
 	CN3UIImage*		m_pBkGndImageRef;		// 배경 이미지 referance (메모리 할당은 children list로 관리)
 	CN3UIImage*		m_pThumbImageRef;		// 가운데 드레그 하여 옮길 수 있는 이미지 referance
 
-	int				m_iMaxPos;									// 최대
-	int				m_iMinPos;									// 최소
-	int				m_iCurPos;									// 현재 값
+	size_t			m_iMaxPos;									// 최대
+	size_t			m_iMinPos;									// 최소
+	size_t 			m_iCurPos;									// 현재 값
 	int				m_iPageSize;								// page단위 이동할때 이동값
 // Operations
 public:
@@ -36,15 +36,15 @@ public:
 	virtual DWORD	MouseProc(DWORD dwFlags, const POINT& ptCur, const POINT& ptOld);
 //	virtual void	Render();
 
-	void			SetRange(int iMin, int iMax);
-	void			SetRangeMax(int iMax) {SetRange(m_iMinPos, iMax);}
-	void			SetRangeMin(int iMin) {SetRange(iMin, m_iMaxPos);}
-	void			SetCurrentPos(int iPos);
-	int				GetPos() const {return m_iCurPos;}
+	void			SetRange(size_t iMin, size_t iMax);
+	void			SetRangeMax(size_t iMax) {SetRange(m_iMinPos, iMax);}
+	void			SetRangeMin(size_t iMin) {SetRange(iMin, m_iMaxPos);}
+	void			SetCurrentPos(size_t iPos);
+	size_t			GetPos() const {return m_iCurPos;}
 	void			SetPageSize(int iSize) {m_iPageSize = iSize;}
 	int				GetPageSize() const {return m_iPageSize;}
-	int				GetMaxPos() const {return m_iMaxPos;}
-	int				GetMinPos() const {return m_iMinPos;}
+	size_t			GetMaxPos() const {return m_iMaxPos;}
+	size_t			GetMinPos() const {return m_iMinPos;}
 protected:
 	void			UpdateThumbPos();							// m_iCurPos를 계산하여 Thumb위치 다시 계산하여 바꾸기
 	void			UpDownThumbPos(int iDiff);					// Thumb위치를 아래 위로 iDiff pixel만큼 움직인 후 m_iCurPos 갱신하기
